Updated Search
I updated the lookUpRecord() and made it more more versatile.
There is a new method that is called searchRecords() that takes it two inputs, fieldType and itemtoSearch. It then uses a switch case to change what is being put into the MySQL query and will print out the query. This reduces the need for repetitive code to print out the result tabel