This is the datase with GUIs
Not all of them are done yet, but the ones that are there work. To make them work you have to run each method separately because the menu isn't integrated yet. 
To run the methods all you have to do is type in which method you want to run in the constructor at the top of the class.